rebind-port: Ignore ACK only packet
Because a packet that contains PATH_CHALLENGE is subject to pacing, an
endpoint might not be able to send it because of pacer.  Meanwhile,
ACK only packet can be sent without pacing limitation as per RFC 9002,
section 7.7.  This means that non-probing ACK only packet may be sent
before sending a packet that contains PATH_CHALLENGE.  To deal with
this situation, ignore ACK only packet when detecting non-probing
packet in rebind-port test.